Adds product cEAN to user's history
