+++ functions/package.json
@@ -15,8 +15,10 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "@types/uuid": "^8.3.0",
     "firebase-admin": "^9.4.1",
-    "firebase-functions": "^3.11.0"
+    "firebase-functions": "^3.11.0",
+    "uuid": "^8.3.1"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^3.9.1",
